Generate cluster information from event logs in Qualification tools
