+++ core/src/main/scala/org/apache/spark/sql/rapids/tool/EventProcessorBase.scala
@@ -206,7 +206,25 @@ abstract class EventProcessorBase[T <: AppBase](app: T) extends SparkListener wi
 
   def doSparkListenerBlockManagerAdded(
       app: T,
-      event: SparkListenerBlockManagerAdded): Unit = {}
+      event: SparkListenerBlockManagerAdded): Unit = {
+    logDebug("Processing event: " + event.getClass)
+    val execExists = app.executorIdToInfo.get(event.blockManagerId.executorId)
+    if (event.blockManagerId.executorId == "driver" && execExists.isEmpty) {
+      // means its not in local mode, skip counting as executor
+    } else {
+      // note that one block manager is for driver as well
+      val exec = app.getOrCreateExecutor(event.blockManagerId.executorId, event.time)
+      exec.hostPort = event.blockManagerId.hostPort
+      event.maxOnHeapMem.foreach { mem =>
+        exec.totalOnHeap = mem
+      }
+      event.maxOffHeapMem.foreach { offHeap =>
+        exec.totalOffHeap = offHeap
+      }
+      exec.isActive = true
+      exec.maxMemory = event.maxMem
+    }
+  }
 
   override def onBlockManagerAdded(blockManagerAdded: SparkListenerBlockManagerAdded): Unit = {
     doSparkListenerBlockManagerAdded(app, blockManagerAdded)
@@ -253,7 +271,16 @@ abstract class EventProcessorBase[T <: AppBase](app: T) extends SparkListener wi
 
   def doSparkListenerExecutorAdded(
       app: T,
-      event: SparkListenerExecutorAdded): Unit = {}
+      event: SparkListenerExecutorAdded): Unit = {
+    logDebug("Processing event: " + event.getClass)
+    val exec = app.getOrCreateExecutor(event.executorId, event.time)
+    exec.host = event.executorInfo.executorHost
+    exec.isActive = true
+    exec.totalCores = event.executorInfo.totalCores
+    val rpId = event.executorInfo.resourceProfileId
+    exec.resources = event.executorInfo.resourcesInfo
+    exec.resourceProfileId = rpId
+  }
 
   override def onExecutorAdded(executorAdded: SparkListenerExecutorAdded): Unit = {
     doSparkListenerExecutorAdded(app, executorAdded)
