+++ core/src/main/scala/org/apache/spark/sql/rapids/tool/profiling/ApplicationInfo.scala
@@ -191,8 +191,6 @@ class ApplicationInfo(
     val index: Int)
   extends AppBase(Some(eLogInfo), Some(hadoopConf)) with Logging {
 
-  // executorId to executor info
-  val executorIdToInfo = new HashMap[String, ExecutorInfoClass]()
   // resourceprofile id to resource profile info
   val resourceProfIdToInfo = new HashMap[Int, ResourceProfileInfoCase]()
 
@@ -227,12 +225,6 @@ class ApplicationInfo(
   override def processEvent(event: SparkListenerEvent) = {
     eventProcessor.processAnyEvent(event)
     false
-  }
-
-  def getOrCreateExecutor(executorId: String, addTime: Long): ExecutorInfoClass = {
-    executorIdToInfo.getOrElseUpdate(executorId, {
-      new ExecutorInfoClass(executorId, addTime)
-    })
   }
 
   // Connects Operators to Stages using AccumulatorIDs
